+++ lldb/test/API/functionalities/data-formatter/data-formatter-stl/generic/map/TestDataFormatterStdMap.py
@@ -2,7 +2,7 @@
 Test lldb data formatter subsystem.
 """
 
-from typing import Optional
+
 import lldb
 from lldbsuite.test.decorators import *
 from lldbsuite.test.lldbtest import *
@@ -11,28 +11,26 @@ from lldbsuite.test import lldbutil
 
 class StdMapDataFormatterTestCase(TestBase):
     def setUp(self):
-        # Call super's setUp().
         TestBase.setUp(self)
-        # Find the line number to break at.
-        self.line = line_number("main.cpp", "// Set break point at this line.")
+        ns = "ndk" if lldbplatformutil.target_is_android() else ""
+        self.namespace = "std"
 
-    @add_test_categories(["libstdcxx"])
-    @expectedFailureAll(bugnumber="llvm.org/pr50861", compiler="gcc")
-    def test_with_run_command(self):
-        self.with_run_command()
+    def check_pair(self, first_value, second_value):
+        pair_children = [
+            ValueCheck(name="first", value=first_value),
+            ValueCheck(name="second", value=second_value),
+        ]
+        return ValueCheck(children=pair_children)
 
-    @add_test_categories(["libstdcxx"])
-    @expectedFailureAll(bugnumber="llvm.org/pr50861", compiler="gcc")
-    def test_with_run_command_debug(self):
-        build_args = {"CXXFLAGS_EXTRAS": "-D_GLIBCXX_DEBUG"}
-        self.with_run_command("__debug::", build_args)
-
-    def with_run_command(self, namespace: str = "", dictionary: Optional[dict] = None):
+    def do_test(self):
         """Test that that file and class static variables display correctly."""
-        self.build(dictionary=dictionary)
         self.runCmd("file " + self.getBuildArtifact("a.out"), CURRENT_EXECUTABLE_SET)
 
-        lldbutil.run_break_set_by_source_regexp(self, "Set break point at this line.")
+        bkpt = self.target().FindBreakpointByID(
+            lldbutil.run_break_set_by_source_regexp(
+                self, "Set break point at this line."
+            )
+        )
 
         self.runCmd("run", RUN_SUCCEEDED)
 
@@ -54,20 +52,24 @@ class StdMapDataFormatterTestCase(TestBase):
         # Execute the cleanup function during test case tear down.
         self.addTearDownHook(cleanup)
 
-        self.runCmd("frame variable ii --show-types")
+        ns = self.namespace
+        self.expect_expr("ii", result_summary="size=0", result_children=[])
 
-        self.runCmd(
-            f'type summary add -x "std::{namespace}map<" --summary-string "map has ${{svar%#}} items" -e'
+        self.expect("frame var ii", substrs=["%s::map" % ns, "size=0", "{}"])
+
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
+
+        self.expect_expr(
+            "ii",
+            result_summary="size=2",
+            result_children=[self.check_pair("0", "0"), self.check_pair("1", "1")],
         )
-
-        self.expect("frame variable ii", substrs=["map has 0 items", "{}"])
-
-        self.runCmd("c")
 
         self.expect(
             "frame variable ii",
             substrs=[
-                "map has 2 items",
+                "%s::map" % ns,
+                "size=2",
                 "[0] = ",
                 "first = 0",
                 "second = 0",
@@ -77,12 +79,13 @@ class StdMapDataFormatterTestCase(TestBase):
             ],
         )
 
-        self.runCmd("c")
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
 
         self.expect(
             "frame variable ii",
             substrs=[
-                "map has 4 items",
+                "%s::map" % ns,
+                "size=4",
                 "[2] = ",
                 "first = 2",
                 "second = 0",
@@ -92,12 +95,13 @@ class StdMapDataFormatterTestCase(TestBase):
             ],
         )
 
-        self.runCmd("c")
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
 
         self.expect(
-            "frame variable ii",
+            "expression ii",
             substrs=[
-                "map has 9 items",
+                "%s::map" % ns,
+                "size=8",
                 "[5] = ",
                 "first = 5",
                 "second = 0",
@@ -108,9 +112,10 @@ class StdMapDataFormatterTestCase(TestBase):
         )
 
         self.expect(
-            "expression ii",
+            "frame var ii",
             substrs=[
-                "map has 9 items",
+                "%s::map" % ns,
+                "size=8",
                 "[5] = ",
                 "first = 5",
                 "second = 0",
@@ -124,7 +129,21 @@ class StdMapDataFormatterTestCase(TestBase):
         self.expect("frame variable ii[0]", substrs=["first = 0", "second = 0"])
         self.expect("frame variable ii[3]", substrs=["first =", "second ="])
 
-        self.expect("frame variable ii[8]", matching=True, substrs=["1234567"])
+        # (Non-)const key/val iterators
+        self.expect_expr(
+            "it",
+            result_children=[
+                ValueCheck(name="first", value="0"),
+                ValueCheck(name="second", value="0"),
+            ],
+        )
+        self.expect_expr(
+            "const_it",
+            result_children=[
+                ValueCheck(name="first", value="0"),
+                ValueCheck(name="second", value="0"),
+            ],
+        )
 
         # check that MightHaveChildren() gets it right
         self.assertTrue(
@@ -140,55 +159,59 @@ class StdMapDataFormatterTestCase(TestBase):
         # self.expect("expression ii[8]", matching=False, error=True,
         #            substrs = ['1234567'])
 
-        self.runCmd("c")
+        self.runCmd("continue")
 
-        self.expect("frame variable ii", substrs=["map has 0 items", "{}"])
+        self.expect("frame variable ii", substrs=["%s::map" % ns, "size=0", "{}"])
 
-        self.runCmd("frame variable si --show-types")
+        self.expect("frame variable si", substrs=["%s::map" % ns, "size=0", "{}"])
 
-        self.expect("frame variable si", substrs=["map has 0 items", "{}"])
-
-        self.runCmd("c")
-
-        self.expect(
-            "frame variable si",
-            substrs=["map has 1 items", "[0] = ", 'first = "zero"', "second = 0"],
-        )
-
-        self.runCmd("c")
+        self.runCmd("continue")
 
         self.expect(
             "frame variable si",
             substrs=[
-                "map has 5 items",
-                '[0] = (first = "four", second = 4)',
-                '[1] = (first = "one", second = 1)',
-                '[2] = (first = "three", second = 3)',
-                '[3] = (first = "two", second = 2)',
-                '[4] = (first = "zero", second = 0)',
+                "%s::map" % ns,
+                "size=1",
+                "[0] = ",
+                'first = "zero"',
+                "second = 0",
+            ],
+        )
+
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
+
+        self.expect(
+            "frame variable si",
+            substrs=[
+                "%s::map" % ns,
+                "size=4",
+                '[0] = (first = "one", second = 1)',
+                '[1] = (first = "three", second = 3)',
+                '[2] = (first = "two", second = 2)',
+                '[3] = (first = "zero", second = 0)',
             ],
         )
 
         self.expect(
             "expression si",
             substrs=[
-                "map has 5 items",
-                '[0] = (first = "four", second = 4)',
-                '[1] = (first = "one", second = 1)',
-                '[2] = (first = "three", second = 3)',
-                '[3] = (first = "two", second = 2)',
-                '[4] = (first = "zero", second = 0)',
+                "%s::map" % ns,
+                "size=4",
+                '[0] = (first = "one", second = 1)',
+                '[1] = (first = "three", second = 3)',
+                '[2] = (first = "two", second = 2)',
+                '[3] = (first = "zero", second = 0)',
             ],
         )
-
-        # check access-by-index
-        self.expect("frame variable si[0]", substrs=["first = ", "four", "second = 4"])
 
         # check that MightHaveChildren() gets it right
         self.assertTrue(
             self.frame().FindVariable("si").MightHaveChildren(),
             "si.MightHaveChildren() says False for non empty!",
         )
+
+        # check access-by-index
+        self.expect("frame variable si[0]", substrs=["first = ", "one", "second = 1"])
 
         # check that the expression parser does not make use of
         # synthetic children instead of running code
@@ -198,20 +221,21 @@ class StdMapDataFormatterTestCase(TestBase):
         # self.expect("expression si[0]", matching=False, error=True,
         #            substrs = ['first = ', 'zero'])
 
-        self.runCmd("c")
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
 
-        self.expect("frame variable si", substrs=["map has 0 items", "{}"])
+        self.expect("frame variable si", substrs=["%s::map" % ns, "size=0", "{}"])
 
-        self.runCmd("frame variable is --show-types")
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
 
-        self.expect("frame variable is", substrs=["map has 0 items", "{}"])
+        self.expect("frame variable is", substrs=["%s::map" % ns, "size=0", "{}"])
 
-        self.runCmd("c")
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
 
         self.expect(
             "frame variable is",
             substrs=[
-                "map has 4 items",
+                "%s::map" % ns,
+                "size=4",
                 '[0] = (first = 1, second = "is")',
                 '[1] = (first = 2, second = "smart")',
                 '[2] = (first = 3, second = "!!!")',
@@ -222,7 +246,8 @@ class StdMapDataFormatterTestCase(TestBase):
         self.expect(
             "expression is",
             substrs=[
-                "map has 4 items",
+                "%s::map" % ns,
+                "size=4",
                 '[0] = (first = 1, second = "is")',
                 '[1] = (first = 2, second = "smart")',
                 '[2] = (first = 3, second = "!!!")',
@@ -230,14 +255,14 @@ class StdMapDataFormatterTestCase(TestBase):
             ],
         )
 
-        # check access-by-index
-        self.expect("frame variable is[0]", substrs=["first = ", "second ="])
-
         # check that MightHaveChildren() gets it right
         self.assertTrue(
             self.frame().FindVariable("is").MightHaveChildren(),
             "is.MightHaveChildren() says False for non empty!",
         )
+
+        # check access-by-index
+        self.expect("frame variable is[0]", substrs=["first = ", "second ="])
 
         # check that the expression parser does not make use of
         # synthetic children instead of running code
@@ -247,46 +272,46 @@ class StdMapDataFormatterTestCase(TestBase):
         # self.expect("expression is[0]", matching=False, error=True,
         #            substrs = ['first = ', 'goofy'])
 
-        self.runCmd("c")
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
 
-        self.expect("frame variable is", substrs=["map has 0 items", "{}"])
+        self.expect("frame variable is", substrs=["%s::map" % ns, "size=0", "{}"])
 
-        self.runCmd("frame variable ss --show-types")
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
 
-        self.expect("frame variable ss", substrs=["map has 0 items", "{}"])
+        self.expect("frame variable ss", substrs=["%s::map" % ns, "size=0", "{}"])
 
-        self.runCmd("c")
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
 
         self.expect(
             "frame variable ss",
             substrs=[
-                "map has 4 items",
-                '[0] = (first = "a Mac..", second = "..is always a Mac!")',
-                '[1] = (first = "casa", second = "house")',
-                '[2] = (first = "ciao", second = "hello")',
-                '[3] = (first = "gatto", second = "cat")',
+                "%s::map" % ns,
+                "size=3",
+                '[0] = (first = "casa", second = "house")',
+                '[1] = (first = "ciao", second = "hello")',
+                '[2] = (first = "gatto", second = "cat")',
             ],
         )
 
         self.expect(
             "expression ss",
             substrs=[
-                "map has 4 items",
-                '[0] = (first = "a Mac..", second = "..is always a Mac!")',
-                '[1] = (first = "casa", second = "house")',
-                '[2] = (first = "ciao", second = "hello")',
-                '[3] = (first = "gatto", second = "cat")',
+                "%s::map" % ns,
+                "size=3",
+                '[0] = (first = "casa", second = "house")',
+                '[1] = (first = "ciao", second = "hello")',
+                '[2] = (first = "gatto", second = "cat")',
             ],
         )
-
-        # check access-by-index
-        self.expect("frame variable ss[3]", substrs=["gatto", "cat"])
 
         # check that MightHaveChildren() gets it right
         self.assertTrue(
             self.frame().FindVariable("ss").MightHaveChildren(),
             "ss.MightHaveChildren() says False for non empty!",
         )
+
+        # check access-by-index
+        self.expect("frame variable ss[2]", substrs=["gatto", "cat"])
 
         # check that the expression parser does not make use of
         # synthetic children instead of running code
@@ -296,6 +321,22 @@ class StdMapDataFormatterTestCase(TestBase):
         # self.expect("expression ss[3]", matching=False, error=True,
         #            substrs = ['gatto'])
 
-        self.runCmd("c")
+        lldbutil.continue_to_breakpoint(self.process(), bkpt)
 
-        self.expect("frame variable ss", substrs=["map has 0 items", "{}"])
+        self.expect("frame variable ss", substrs=["%s::map" % ns, "size=0", "{}"])
+
+    @add_test_categories(["libc++"])
+    def test_libcxx(self):
+        self.build(dictionary={"USE_LIBCPP" : 1})
+        self.do_test()
+
+    @add_test_categories(["libstdcxx"])
+    def test_libstdcxx(self):
+        self.build(dictionary={"USE_LIBSTDCPP" : 1})
+        self.do_test()
+
+    @add_test_categories(["libstdcxx"])
+    def test_libstdcxx_debug(self):
+        self.build(dictionary={"USE_LIBSTDCPP" : 1,
+                               "CXXFLAGS_EXTRAS": "-D_GLIBCXX_DEBUG"})
+        self.do_test()
